requirements.txt
# YOLOv8 Training Requirements
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.23.0
opencv-python>=4.7.0
matplotlib>=3.7.0
